refactor(desktop): isolate proactive QA diagnostics
