--- desktop/macos/changelog/unreleased/20260813-proactive-qa-diagnostics.json
@@ -1,3 +1,0 @@
-{
-  "change": "Improved proactive-assistant QA diagnostics for capture gates and bounded service errors"
-}
